Move FluxModelLoaderInvocation to its own file. model.py was getting bloated.
